feat: :sparkles: More styles are added to improve the UI/UX for the user, making the design more attractive and encouraging the client to make a purchase on the page.
More CSS styles are added to give the page robustness and customization, aiming to provide the user with a better experience and smoother interaction. Additionally, everything is carefully structured and meticulously planned so that the user spends more time on the page and is encouraged to purchase as many products as possible.

by@Alexis-Blanco